[refactor] (Decomp of #5251 1/n) Move ParallelExecutor out of async engine (#5351)
* [refactor] (Decomp of #5251 1/n) Move ParallelExecutor out of async engine

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci
